code/v2/app/dev/requirements.txt
Flask==3.0.0
mmdet==3.2.0
pandas==2.0.3
Pillow==10.0.1
Pillow==10.1.0
pyarrow==13.0.0
